#993. FFI tests refactored and improved

+++ LibTest/ffi/Int8/Int8_A01_t02.dart
@@ -14,7 +14,7 @@ import "../../../Utils/expect.dart";
 
 void main() {
   Pointer<Int8> p1 = calloc<Int8>();
-  Pointer<Int16> p2 = new Pointer<Int16>.fromAddress(p1.address);
+  Pointer<Uint8> p2 = new Pointer<Uint8>.fromAddress(p1.address);
   try {
     p2.value = 42;
     Expect.equals(42, p1.value);
